refactor:renamed internal package
