+++ internalpackage/services/factory.go
@@ -2,14 +2,14 @@ package services
 
 import (
 	"fmt"
-	"github.com/Tradia-HQ/tradia-file-uploader-service/internal/config"
-	"github.com/Tradia-HQ/tradia-file-uploader-service/internal/services/digitalOcean"
-	"github.com/Tradia-HQ/tradia-file-uploader-service/internal/services/gcs"
-	"github.com/Tradia-HQ/tradia-file-uploader-service/internal/services/s3"
-	"github.com/Tradia-HQ/tradia-file-uploader-service/pkg"
+	"github.com/Tradia-HQ/tradia-file-uploader-service/internalpackage/config"
+	"github.com/Tradia-HQ/tradia-file-uploader-service/internalpackage/services/digitalOcean"
+	"github.com/Tradia-HQ/tradia-file-uploader-service/internalpackage/services/gcs"
+	"github.com/Tradia-HQ/tradia-file-uploader-service/internalpackage/services/interfaces"
+	"github.com/Tradia-HQ/tradia-file-uploader-service/internalpackage/services/s3"
 )
 
-func NewUploader(cfg *config.Config) (pkg.FileUploader, error) {
+func NewUploader(cfg *config.Config) (interfaces.FileUploader, error) {
 	switch cfg.Provider {
 	case "gcs":
 		return gcs.NewGCSService(cfg.GCS.ProjectID, cfg.GCS.BucketName, cfg.GCS.CredentialsPath)
