+++ internalpackage/services/interfaces/uploader.go
@@ -1,4 +1,4 @@
-package pkg
+package interfaces
 
 import (
 	"context"
